Move Principal interface and ResourcePrincipal interface to common/api.
