--- src/main/kotlin/org/wfanet/measurement/reporting/service/api/BUILD.bazel
@@ -1,16 +1,0 @@
-load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
-
-package(default_visibility = [
-    "//src/main/kotlin/org/wfanet/measurement:__subpackages__",
-    "//src/test/kotlin/org/wfanet/measurement:__subpackages__",
-])
-
-kt_jvm_library(
-    name = "context_keys",
-    srcs = ["ContextKeys.kt"],
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal",
-        "@wfa_common_jvm//imports/java/io/grpc:api",
-        "@wfa_common_jvm//imports/java/io/grpc:context",
-    ],
-)
