+++ src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha/BUILD.bazel
@@ -16,8 +16,8 @@ kt_jvm_library(
     srcs = ["ReportingSetsService.kt"],
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/api:public_api_version",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal_server_interceptor",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:reporting_principal",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:resource_key",
         "//src/main/proto/wfa/measurement/api/v2alpha:page_token_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/reporting:reporting_sets_service_kt_jvm_grpc_proto",
@@ -36,8 +36,8 @@ kt_jvm_library(
         ":resource_key",
         "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:event_group_metadata_parser",
         "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/eventfiltration:event_filters",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal_server_interceptor",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:reporting_principal",
         "//src/main/proto/wfa/measurement/api/v2alpha:event_group_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:event_group_metadata_descriptors_service_kt_jvm_grpc_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:event_groups_service_kt_jvm_grpc_proto",
@@ -55,8 +55,8 @@ kt_jvm_library(
         "//src/main/kotlin/org/wfanet/measurement/api:api_key_constants",
         "//src/main/kotlin/org/wfanet/measurement/api:public_api_version",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:encryption_key_pair_store",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal_server_interceptor",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:reporting_principal",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:resource_key",
         "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:set_operation_compiler",
         "//src/main/proto/wfa/measurement/api/v2alpha:certificate_kt_jvm_proto",
@@ -97,10 +97,22 @@ kt_jvm_library(
 )
 
 kt_jvm_library(
-    name = "principal",
-    srcs = ["Principal.kt"],
+    name = "context_keys",
+    srcs = ["ContextKeys.kt"],
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/common/api:principal",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:reporting_principal",
+        "@wfa_common_jvm//imports/java/io/grpc:api",
+        "@wfa_common_jvm//imports/java/io/grpc:context",
+    ],
+)
+
+kt_jvm_library(
+    name = "reporting_principal",
+    srcs = ["ReportingPrincipal.kt"],
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:resource_key",
+        "//src/main/kotlin/org/wfanet/measurement/common/api:principal",
     ],
 )
 
@@ -109,8 +121,8 @@ kt_jvm_library(
     srcs = ["PrincipalServerInterceptor.kt"],
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/common/identity",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api:context_keys",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:principal",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:context_keys",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha:reporting_principal",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
         "@wfa_common_jvm//imports/java/io/grpc:api",
         "@wfa_common_jvm//imports/java/io/grpc:context",
