+++ src/main/kotlin/org/wfanet/measurement/reporting/service/api/v1alpha/ContextKeys.kt
@@ -1,0 +1,22 @@
+// Copyright 2022 The Cross-Media Measurement Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.wfanet.measurement.reporting.service.api.v1alpha
+
+import io.grpc.Context
+
+object ContextKeys {
+  /** This is the context key for the authenticated [ReportingPrincipal]. */
+  val PRINCIPAL_CONTEXT_KEY: Context.Key<ReportingPrincipal> = Context.key("principal")
+}
